Atualização

+++ 1 Variáveis/5_5_calculadora.py
@@ -1,4 +1,4 @@
-num1 = float(input('Digite um número por favor: '))
-num2 = float(input('Outro número: '))
+num1 = int(input('Digite um número por favor: '))
+num2 = int(input('Outro número: '))
 
 print(f'A soma dos número é de {num1 + num2}')
